backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1
rq==1.15.1
qdrant-client==1.6.9
minio==7.2.0
spacy==3.7.2
transformers==4.35.2
torch==2.1.1
Pillow==10.1.0
pytesseract==0.3.10
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
alembic==1.12.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
httpx==0.25.2
aiofiles==23.2.0
numpy==1.24.4
pandas==2.1.3
